+++ src/main/java/base/designpatterns/DesignPattern.java
@@ -1,4 +1,4 @@
-package base;
+package base.designpatterns;
 
 /**
  * Instantiation can be done according to the following example. Using curly brackets to enter an
