+++ src/main/java/base/designpatterns/Pattern.java
@@ -1,4 +1,4 @@
-package base;
+package base.designpatterns;
 
 /**
  * An enum that contains all patterns we are able to identify and verify. If a new pattern has been
