+++ src/main/java/base/designpatterns/verifiers/singleclassverifiers/IPatternVerifier.java
@@ -1,6 +1,8 @@
-package patternverifiers;
+package base.designpatterns.verifiers.singleclassverifiers;
 
 import com.github.javaparser.ast.CompilationUnit;
+
+import base.util.Feedback;
 
 /**
  * An interface that defines the obligatory behaviour that all pattern verifiers must implement.
